throw exception if directory iteration fails before attempting delete

+++ denkbares-Utils/src/main/java/com/denkbares/util/nio/Paths.java
@@ -234,6 +234,8 @@ public class Paths {
 			@Override
 			@NotNull
 			public FileVisitResult postVisitDirectory(Path dir, IOException exc) throws IOException {
+				// if iterating the directory failed, deleting it would only mask the original error
+				if (exc != null) throw exc;
 				Files.delete(dir);
 				return FileVisitResult.CONTINUE;
 			}
